refactor: kafka common ops
